impl variation picking fallback
